requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
requests==2.32.3
pandas== 2.2.3
python-dotenv==1.0.1
pymysql==1.1.1
sqlalchemy==2.0.37
aiohttp==3.11.12
cryptography